+++ tests/testthat/test-pivot.R
@@ -1,56 +1,6 @@
-# mini_pivot_wider ----
+# multi_id_pivot_wider ----
 
-test_that("mini_pivot_wider works as expected.", {
-  test_data <- data.frame(
-    the_obs = c("A", "A", "A", "B", "B", "B", "C", "D"),
-    the_obs2 = c("Ax", "Ax", "Ax", "Bx", "Bx", "Bx", "Cx", "Dx"),
-    the_param = c("weight", "height", "gender", "weight", "gender", "height", "height", "precondition"),
-    the_val = c(65, 165, "M", 66, "F", 166, 155, TRUE)
-  )
-
-  expected_data <- data.frame(
-    id = c("A", "B", "C", "D"),
-    gender = c("M", "F", NA, NA),
-    height = c("165", "166", "155", NA),
-    precondition = c(NA, NA, NA, "TRUE"),
-    weight = c("65", "66", NA, NA)
-  )
-
-  res <- expect_silent(mini_pivot_wider(test_data, "the_obs", "the_param", "the_val"))
-  expect_identical(expected_data, res)
-})
-
-test_that("mini_pivot_wider returns an error when required column are not present.", {
-  test_data <- data.frame(
-    the_obs = c("A", "A", "A", "B", "B", "B", "C", "D"),
-    the_obs2 = c("Ax", "Ax", "Ax", "Bx", "Bx", "Bx", "Cx", "Dx"),
-    the_param = c("weight", "height", "gender", "weight", "gender", "height", "height", "precondition"),
-    the_val = c(65, 165, "M", 66, "F", 166, 155, TRUE)
-  )
-
-  expect_error(mini_pivot_wider(
-    test_data,
-    "Must be a subset of {'the_obs','the_obs2','the_param','the_val'}, but is {'the_obs2','the_param2','the_val2'}."
-  ))
-})
-
-test_that("mini_pivot_wider returns an error when data are duplicated.", {
-  test_data <- data.frame(
-    the_obs = c("A", "A", "A", "B", "B", "B", "C", "D"),
-    the_obs2 = c("Ax", "Ax", "Ax", "Bx", "Bx", "Bx", "Cx", "Dx"),
-    the_param = c("weight", "weight", "gender", "weight", "gender", "height", "height", "precondition"),
-    the_val = c(65, 65, "M", 66, "F", 166, 155, TRUE)
-  )
-
-  expect_error(mini_pivot_wider(
-    test_data,
-    "Assertion on 'any(duplicated(data[, c(obs_id, param_from)]))' failed: Must be FALSE."
-  ))
-})
-
-# multi_pivot_wider ----
-
-test_that("multi_pivot_wider works as expected.", {
+test_that("multi_id_pivot_wider works as expected.", {
   test_data <- data.frame(
     the_obs = c("A", "A", "A", "B", "B", "B", "C", "D"),
     the_obs2 = c("Ax", "Ax", "Ax", "Bx", "Bx", "Bx", "Cx", "Dx"),
@@ -67,11 +17,59 @@ test_that("multi_pivot_wider works as expected.", {
     weight = c("65", "66", NA, NA)
   )
 
-  res <- expect_silent(multi_pivot_wider(test_data, c("the_obs", "the_obs2"), "the_param", "the_val"))
-  expect_identical(expected_data, res)
+  res <- expect_silent(multi_id_pivot_wider(test_data, c("the_obs", "the_obs2"), "the_param", "the_val"))
+  expect_identical(res, expected_data)
 })
 
-test_that("multi_pivot_wider works as expected with drop_na argument.", {
+test_that("multi_id_pivot_wider works as expected with factor as value.", {
+  test_data <- data.frame(
+    the_obs = c("A", "A", "A", "B", "B", "B", "C", "D"),
+    the_obs2 = c("Ax", "Ax", "Ax", "Bx", "Bx", "Bx", "Cx", "Dx"),
+    the_param = c("weight", "height", "gender", "weight", "gender", "height", "height", "precondition"),
+    the_val = factor(c(65, 165, "M", 66, "F", 166, 155, NA))
+  )
+
+  expected_data <- data.frame(
+    the_obs = c("A", "B", "C", "D"),
+    the_obs2 = c("Ax", "Bx", "Cx", "Dx"),
+    gender = factor(c("M", "F", NA, NA), c("155", "165", "166", "65", "66", "F", "M")),
+    height = factor(c("165", "166", "155", NA), c("155", "165", "166", "65", "66", "F", "M")),
+    precondition = factor(c(NA, NA, NA, NA), c("155", "165", "166", "65", "66", "F", "M")),
+    weight = factor(c("65", "66", NA, NA), c("155", "165", "166", "65", "66", "F", "M"))
+  )
+
+  res <- expect_silent(multi_id_pivot_wider(test_data, c("the_obs", "the_obs2"), "the_param", "the_val"))
+  expect_identical(res, expected_data)
+})
+
+test_that("multi_id_pivot_wider works as expected with factor as value and drop_lvl = TRUE", {
+  test_data <- data.frame(
+    the_obs = c("A", "A", "A", "B", "B", "B", "C", "D"),
+    the_obs2 = c("Ax", "Ax", "Ax", "Bx", "Bx", "Bx", "Cx", "Dx"),
+    the_param = c("weight", "height", "gender", "weight", "gender", "height", "height", "precondition"),
+    the_val = factor(c(65, 165, "M", 66, "F", 166, 155, NA))
+  )
+
+  expected_data <- data.frame(
+    the_obs = c("A", "B", "C", "D"),
+    the_obs2 = c("Ax", "Bx", "Cx", "Dx"),
+    gender = factor(c("M", "F", NA, NA), c("F", "M")),
+    height = factor(c("165", "166", "155", NA), c("155", "165", "166")),
+    precondition = factor(c(NA, NA, NA, NA)),
+    weight = factor(c("65", "66", NA, NA), c("65", "66"))
+  )
+
+  res <- expect_silent(
+    multi_id_pivot_wider(
+      test_data,
+      c("the_obs", "the_obs2"), "the_param", "the_val",
+      drop_lvl = TRUE
+    )
+  )
+  expect_identical(res, expected_data)
+})
+
+test_that("multi_id_pivot_wider works as expected with drop_na argument.", {
   test_data <- data.frame(
     the_obs = c("A", "A", "A", "B", "B", "B", "C", "D"),
     the_obs2 = c("Ax", "Ax", "Ax", "Bx", "Bx", "Bx", "Cx", "Dx"),
@@ -87,11 +85,13 @@ test_that("multi_pivot_wider works as expected with drop_na argument.", {
     weight = c("65", "66", NA, NA)
   )
 
-  res <- expect_silent(multi_pivot_wider(test_data, c("the_obs", "the_obs2"), "the_param", "the_val", drop_na = TRUE))
+  res <- expect_silent(
+    multi_id_pivot_wider(test_data, c("the_obs", "the_obs2"), "the_param", "the_val", drop_na = TRUE)
+  )
   expect_identical(expected_data, res)
 })
 
-test_that("multi_pivot_wider works as expected with drop_na argument.", {
+test_that("multi_id_pivot_wider works as expected with drop_na argument.", {
   test_data <- data.frame(
     the_obs = c("A", "A", "A", "B", "B", "B", "C", "D"),
     the_obs2 = c("Ax", "Ax", "Ax", "Bx", "Bx", "Bx", "Cx", "Dx"),
@@ -107,11 +107,13 @@ test_that("multi_pivot_wider works as expected with drop_na argument.", {
     weight = c("65", "66", NA, NA)
   )
 
-  res <- expect_silent(multi_pivot_wider(test_data, c("the_obs", "the_obs2"), "the_param", "the_val", drop_na = TRUE))
+  res <- expect_silent(
+    multi_id_pivot_wider(test_data, c("the_obs", "the_obs2"), "the_param", "the_val", drop_na = TRUE)
+  )
   expect_identical(expected_data, res)
 })
 
-test_that("multi_pivot_wider works as expected when the unique identification deoends on several columns", {
+test_that("multi_id_pivot_wider works as expected when the unique identification depends on several columns", {
   test_data <- data.frame(
     the_obs = c("A", "A", "B"),
     the_obs2 = c("Ax2", "Ax2", "Ax"),
@@ -119,7 +121,7 @@ test_that("multi_pivot_wider works as expected when the unique identification de
     the_val = c(65, 65, 65)
   )
 
-  expect_error(multi_pivot_wider(test_data, c("the_obs", "the_obs2"), "the_param", "the_val"))
+  expect_error(multi_id_pivot_wider(test_data, c("the_obs", "the_obs2"), "the_param", "the_val"))
 
   test_data2 <- data.frame(
     the_obs = c("A", "A", "B"),
@@ -134,7 +136,7 @@ test_that("multi_pivot_wider works as expected when the unique identification de
     weight = c(65, 65, 65)
   )
 
-  res <- expect_silent(multi_pivot_wider(test_data2, c("the_obs", "the_obs2"), "the_param", "the_val"))
+  res <- expect_silent(multi_id_pivot_wider(test_data2, c("the_obs", "the_obs2"), "the_param", "the_val"))
   expect_identical(res, expected_data)
 })