+++ NEWS.md
@@ -1,5 +1,8 @@
 # dunlin 0.1.5.9000
 
+* `join_adsub_adsl` now sends a warning when the continuous or categorical variables are missing or all `NA`. All `NA` variables can now be kept using `drop_na = FALSE`. In addition, missing levels in the pivoted columns can be dropped with `drop_lvl = TRUE`
+* Remove `mini_pivot_wider` function.
+
 # dunlin 0.1.5
 
 * `empty_rule` is removed now. `rule()` will create a normal `rule` object.